Finished mobile API

+++ WebAPI/Models/App/Models/ProgressInClass.cs
@@ -10,7 +10,7 @@ namespace WebAPI.Models.App
         public int Id { get; set; }
         public int StudentId { get; set; }
         public int ClassId { get; set; }
-        public float Result { get; set; }
+        public double Result { get; set; }
         public int FinishedLessonNum { get; set; }
     }
 }
